Crate context_switch

This is a wrapper crate around all other context switch implementation crates that helps to manage the complex configuration options involving SIMD and personalities.

Structs

Context

The registers saved before a context switch and restored after a context switch for SSE-enabled Tasks.

Functions

context_switch⚠

Switches context from an SSE Task to another SSE Task.

read_first_register

Reads the value of the first register from the actual CPU register hardware.